+++ plugins/edulume-core/src/Infrastructure/Wp/Plugin.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Edulume\Core\Infrastructure\Wp;
 
+use Edulume\Core\Infrastructure\Content\ContentRegistrar;
 use Edulume\Core\Infrastructure\Theming\StylesheetEnqueuer;
 
 /**
@@ -49,6 +50,7 @@ final class Plugin
         add_action('init', [$this, 'loadTextDomain']);
 
         $this->stylesheetEnqueuer()->register();
+        (new ContentRegistrar())->register();
     }
 
     public function activate(): null